Initial draft of "Components in A Graph".
The question is really poorly explained, and points to a non-existent
diagram "for more clarity" (sigh), so I'm not sure if this is correct or
not.  Give the correct answer for the sample testcase (which has 0
explanation for it), at least.

Unsuccessful submission of "Components in a Graph" - segmentation
faults.  Possibly stack overflow?