Added a test method "blah" which basically counts the number of node id's at this subtree.
This uses unique_ptrs of trees to avoid any chance of accidentally
copying instead of moving maps.

This is still glacially slow - I've clearly messed up some performance
analysis, but have no idea what it could be.

It could be that the whole "disjoint set" thing is logically wrong, but
I thought that was what would be needed for e.g.

https://www.hackerrank.com/challenges/coloring-tree/problem

which is what "blah" is supposed to be solving.  Hmmm.