Avoid repeated elements in toProcess by adding a "visited" tracker.
No longer TLE, but now wrong answer XD

UPDATE: Oh, ffs - the problem statement says:

"you must remove exactly 1 digit from A and append it to X (to the right
side of X)"

I thought it meant "append the digit", but no - it means append *the
copy of A after you removed the digit".

Good god.