Add some memoisation in, but it's not enough - Number Games still too slow.
I think it's because we're continually re-adding things to process, and
we should be storing, for each "thing", whether it can lead to 0 or not,
and making use of this to avoid re-processing.

Not quite sure how to do that with the "BFS" approach - would a DFS
work?

NB: hoisting "processed" and "visited" up one level out of the for...loop
gives the wrong answer.