Optimised solution for "Dilemma" based on empirical observation.
It seems that it is a win if and only if there is an odd number of cards
facing up in the initial configuration - no proof of  this, yet.

Possibly a proof by induction on number of cards would work - initial
move splits into two subgames of strictly smaller size.