Some exploratory work for Encoding.
Looking promising for finding a recurrence relation for finding the sum
of f(x) over all numbers with len digits beginning with digit d
(sumOfFForNumDigitsBeginningWith[len][d]).

This is the core of what I envisage to be the solution - the rest is
very fiddly and error-prone, but fundamentally "easy", I think.

Actually got the recurrence right first time, bizarrely.