Breakthrough with CaGR!
Finally occurred to me last night some time after midnight: for each a2,
I'm laboriously constructing, over several different passes:

- the set of nodes that are less/ greater than a2 that are reached
before a2 in a DFS with the original child order (but which are not
ancestors of a2);

- the set of nodes that are less/ greater than a2 that are reached
before a2 in a DFS with the *reversed* child order (but, again, which are not
ancestors of a2); and

- the set of ancestors of nodes that are less/ greater than a2.

Taking the union of all three sets - which again, I'm laboriously doing
over multiple passes - gives simply the set of nodes which are less/
greater than a2 *which are not descendents of a2*, and, since we already
know the set of nodes less/ greater  than a2 which *are* descendents of
a2, and the total number of nodes which are less/ greater than a2 full
stop, we can easily calculate the total of the union of the three things
above.

Implement code that does just this; passes randomised tests.