Doh - the min number of moves is for N, not M!
Re-submitted, and got lots of REs - turned out that I have
verifySolution on, but that shouldn't cause any problems, I don't think
... ?

Oh - unless the result is -1, in which case it will try to verify it.
Hmm.

Edit: Wait - this can't possibly be right - it must have been correct
the first time (with M instead of N).  Surely?