Begin work on a possible new problem of my own: "heavy bits".
Just brute force for the time being.

I have no idea if there is an efficient solution to this, but I suspect
there is an O(N) one.

Problem statement: given a binary string B (i.e. a string where each
character is either a '1' or a '0'), let DS(B) be the set of all
*distinct* substrings of B.

For a binary string b, let w(b) be max(# 1's in b, # 0's in b).

Compute sum [b in DS(B)] { weight(b) }.