More (performance) diagnostics for CONTREE, plus allow toggling between CD+Old approaches with a #define.
Have figured out the reason for discrepancy in performance: consider the
amount of nodes processed if the tree is simply a long chain of nodes.

With CD, it's O(N log N); with the old approach, it is O(N).  In
general, with "deep" trees, we should expect the old approach to do
better.

Interestingly, it seems to do better with "shallow", "fat" trees, too.