Give TestFiles consecutive numberings which span *across* subtasks.
i.e. don't restart the testFileNum counter when we change to a new subtask.

To keep the TestFile's listed number in sync with its filename, we *first* group
testfiles by subtask, then by added order in the TestSuite.  We use this
same ordering when assigning the TestFile its filename and when printing
out its number.