Get some interesting testcases for CONTREE subtask 1.
Just used trial and error (and a modified
counters-on-a-tree-ultra-brute-force.cpp which outputs the number
of game states explored and the number of wins for Bob) to come up
with seed values that give a decent amount of states to explore and an
"interesting" number of Wins for Bob.

Also increased maxNumCountersOverAllNodes and maxNumCountersPerNode from 5 to 7 - 5 is too small to have anything interesting!