Begin working on a "cheat" solution to Equilateral Treeangles.
This is a simpler algorithm than the "official" one - I want to see if
there are any testcases (currently existing or otherwise) that can break
it (i.e. by triggering a TLE)

Currently incomplete - there's an optimisation missing, so it currently
TLE'ing on most testcases.

Currently giving incorrect results, so there's a micalculation
somewhere!