REFACTOR,EQUITREE,TG: Begin to move stuff into new helper.
The new helper is makeSquatGraphWhereAllNodesHaveDegreeAtLeast3.
I'm doing this piecemeal as the quality of the testcase (as measured by
how badly it makes the EQUITREE cheat implementation timeout) seems to
be behaving in a way that is very surprising: moving the "add 50'000
neighbours to the rootNode" code to after the "squat graph" has been
created seems to make the testcase *worse* i.e. it is easier to cheat
with.

I have no idea why this is the case - it should be the opposite way
round!