COUNTREE,TGC: Update subtask1 constraints to allow sample testcase XD
The sample testcase has a total of 10 counters, but we originally had a
limit of a total of 6 counters for subtask1 - increase it to 10, and
decrease the number of nodes to 10 to compensate.

This required me to re-generate the "interestingSeeds" for the
"reasonably large statespace" subtask1 testfile that is designed to be
passable by a naive implementation.