MOVCOINS2,tgc: Begin work on findBobWinningRelocatedHeightsForNodes util
Unfortunately, this now fails to link: it seems that test-lib.h is *not*
properly designed to be used as a header-only file, as including it in
multiple .cpp files, as we are implicitly doing here with
move-the-coins-2-testcase-generator.cpp and utils.cpp, causes linker
errors (none of the method definitions are declared "inline", etc).

So we'll have to ditch the whole "utils.cpp" and probably bung
everything into utils.h.