MOVCOINS2,TGC: Initial implementation of findGrundyNumberIfRelocatedNode
Needs testing!

This is the brute-force version (takes O(N)): using this would make the
problem of finding, for each node, the heights that it can be
re-parented to to give a Bob Win O(N^3), which would be appallingly
slow.

It will eventually be replaced with an implementation based on "Counters
on a Tree"'s DistTracker class.