Snapshot of choicesWithRemovals implementation; incomplete
This will simulate "choose c_1th of the initial items and remove it;
c_2nd of the remaining items and remove it ... c_ith of the remaining
items and remove it" etc without actually performing the removals i.e.
each c_i is adjusted to point to the item in the initial items that
would be chosen if we *had* removed the previous i-1 chosen elements.

If M is the number of elements we can choose from, and N is the number
of items to choose & remove, then the (currently incomplete) optimised
version will run in O(N log N) i.e. independent of M.