When we insertValue in a Persistent AVLTree, wipe all subsequent roots
If, starting from a default-constructed AVLTree, we insertValue 100
times, then switchToRevision 5, then insertValue again, all the
m_rootForRevision for revisions 6, 7, ... are destroyed and replaced
with the single new root for revision resulting from this call to
insertValue.

This is more in-line with how an Undo-stack works when we undo a few
times then perform an action - all actions to the right of the current
point in the Undo stack are deleted and replaced with just this new
action.