REFACTOR: In AVLTree::rotateLeft/Right, use better names than newNewSubtreeRoot!
This newNewSubtreeRoots are created as *replacements* for
newSubtreeRoot, as persistence requires that we replace Nodes rather
than change their properties, so "replacementNewSubtreeRoot" is a better
name - use that!

Similarly, "newSubtreeRoot" -> "replacementSubtreeRoot" - this was even
worse originally, as the name "newSubtreeRoot" was already being used to
refer to some completely different concept in the same function - yuck!