Begin work on findKthFromPair for AVLTree
As the name suggests, this has the effect of merging the two lists of
(sorted) numbers represented by the two AVLTrees, and finding the Kth
number in the resulting (sorted) list, except it occurs in O(log2(size
of first tree) * log2(size of second tree)).

Assumes that the resulting list has no repetitions.

Currently not giving the correct answer.