Begin work on new Problem of my own: Simple Markdown Editor
Vaguely inspired by https://www.hackerrank.com/challenges/simple-text-editor,
which unfortunately had very weak testcases.

This is a more advanced Problem, though, requiring the use of a Persistent
Balance Tree (or some other appropriate structure).

The basic Problem: you have a simplified view of a Markdown Editor - you
don't care about the specific characters in the Document - just whether
they are _formatting_ (represented by '*') or not (represented by 'X').
You also have an Undo Stack (initially empty) and an Undo Stack Pointer
(initially 0).

You must process queries online (the presence of the - as yet unused -
decryptionKey will eventually enforce this); the queries are:

  1. F <position> - insert a _formatting character_ ('*') at the given
     position in the document
  2. N <position> - insert a _not formatting character_ ('X') at the
     given position in the document
  3. Q <position> - determine whether the given position in the document
     is part of a _formatted range_ (see later) and, if so, the size of
     that range
  4. U <numToUndo> - undo the last numToUndo insertion operations in
     the Undo stack
  5. R <numToRedo> - redo the last numToUndo insertion operations in the
     Undo stack

When processing either of the insertion operations (F or N), we erase
all elements in the Undo Stack to right of the Undo Stack Pointer (if
any); append a description of the insertion operation to the Undo Stack;
and increment the Undo Stack Pointer.

To determine the _formatted ranges_ of the Document, we _parse_ the
Document as follows:

  1. Find all consecutive pairs of formatting characters from left to right
     i.e. pair the first and second formatting characters; then the 3rd and
     fourth; etc.
  2. The formatted ranges are the ranges of positions between each pair
     of formatting characters we formed in 1.

Currently, just has a Brute Force implementation, which has a bunch of missing
features (no Undo/ Redo support; no decryptionKey usage ; etc).

The optimised implementation should be simple given the Persistent
AVLTree I've developed, but I haven't started working on it, yet.