SIMKDOWN: Create a solveBruteForce function and Query struct to pass to it
Gather up all the input into a vector<Query> and pass it to
solveBruteForce rather than having the main-loop read input and
conditionally pass it to the brute-force stuff for processing.

This will give a cleaner separation between Brute-Force and Optimised
solutions when we come to work on the latter, and enable us to compare
their answers more easily.