SIMKDOWN: Fix yet more bugs
Address the TODO in adjustRunToLeftOfNodeToRightOf - now the subTreeRoot
is always COW'd, rather than spraying COWs all over the place.

This means we can now just call updateInfoFromChildren at the end of
adjustRunToLeftOfNodeToRightOf, which fixes errors with node position
calculation (in particular, the sumOfDescendantValues for nodes above
the node we end up adjusting are now correct).

The optimised solution now gives the same results as the brute force one
for the sample testcase.