SIMKDOWN: Fix bug in adjustRunToLeftOfNodeToRightOf
The maxPosInLeftSubchild calculation was off by one (still including
the contribution of currentNode), so adjustRunToLeftOfNodeToRightOf
would sometimes erroneously explore the left subchild thinking that the
node to adjust is in there, when it wasn't - in this situation, no node
would end up being adjusted.

Now passing loads of random testcases - might actually be correct, now!