Rework BSTOPS so that delete'ing a node doesn't delete whole subtree
It was stupid of me to assume that deleting a node meant deleting the
whole subtree rooted at the node, and I have no idea why I did it XD

There are many ways of maintaining a BST when deleting a node with two
children (i.e. the Problem Statement is ambiguous); I've opted for what
looks like the logic from this article:

https://www.geeksforgeeks.org/binary-search-tree-set-2-delete/

Needs debugging, though :)

Abandoned unique_ptr; it's more trouble than it's worth in this Problem.