BSTOPS - stop trying to track parents; other magic to get AC
Can't be arsed to write a proper description of this, because frankly,
there is no logic to it.

The only thing I can think of here is that the Problem Setter had a very
specific implementation in mind with very arbitrary rules, and generated
the testcases based on the output of his implementation.

Evidence: doing it the way I've done it here prints out the correct(!)
*two* positions when a node with two children gets deleted.  There is
nothing in the Problem Statement to suggest that any more than one
should be printed.  Doing it this "dumb" way does precisely that,
though.