MOVCOIN2,TGC: Add a representative dummy testcase for performance testing
This is roughly a "worst-case": the max height of a node is approx 90'000,
and any node can be reparent to a new parent node with any height up to
this max height.

This testcase took approx 1 hr 25 mins on my desktop to find the set of
Bob Wins.

NB: There are still no "proper" testcases for MOVCOIN2.