Fix bug in chooseKRandomIndicesFrom that returned out-of-bounds indices
One of the returned indices could be exactly equal to numToChooseFrom,
which is an error - the cause was a misuse of rnd.next - I used

    rnd.next(0, numRemaining)

which returns a random number in the range [0, numRemaining], instead of
[0, numRemaining), as desired.

The fix consists of just using

    rnd.next(numRemaining)

which returns a random number in the desired range.

Add an assertion so that this doesn't happen again.