MOVCOIN2,TGC: Fix another source of non-determinacy
The order in which arguments to a function are evaluated is undefined.
Thus, when we originally did e.g.:

addQueriesAlongFirstHalfOfChain(queries, arm1, rnd.next(43'000, 47'000), rnd.next(20.0, 30.0), nodesAtHeight);

the order in which the two rnd.next(...) expressions were evaluated was
undefined, so it would be possible for them to be called in the opposite
order and thus give entirely different results.

Refactor to remove this possiblility by calculating numToGenerate and
percentageBobWin deterministically before the call to
addQueriesAlongFirstHalfOfChain, and passing these computed values to
addQueriesAlongFirstHalfOfChain.