MOVCOIN2,TGC: Finish subtask3.maxNumTestcases test
We need to ensure that the number of nodes is > 1, else no queries can
be computed - which causes a bit of difficulty since numQueries > 0!

Also, suspend caching (only for this test) - looks like the treeHash function isn't up to the
task, and we get occasional cache collisions if we have many small trees.