MOVCOIN2,TGC: Obscure a possibly-detectable pattern in the testcases
An Alice-Win's nodeToReparent is significantly more likely, by choice of the
original chosenBobWinPairs, to be part of more than one query whose
newParentHeight's are the same.

During the "add small height offset" query of Alice-Win generation in
generateQueriesFromNodes, add the possibility of adding a new NodeAndHeight
which is a complete copy of an existing NodeAndHeight, increasing the
probabilty that a *Bob*-Win's nodeToReparent will be part of more than one
query whose newParentHeight's are the same.