MOVCOIN2,PS: The images are all too large; experiment with shrinking one
The many of the example images do not fit in the Problem Statement
content area (I didn't notice this at first, as the Problem Statement as
displayed in the preview do not have the vertical margins that are
present in the "real" Problem Statement page).

Experiment with resizing all the text, then exporting as a
proportionately smaller PNG.

MOVCOIN2-EX1-4of4.odg/.png are the guinea pig for this, and seem to work
quite well: will apply the same changes to the other .odg/.pngs in due
course.