MVCN2TST: Begin working on the bit that finds the height of newParent
Just verifying a hypothesis at the moment - I'm actually slightly
confused as to why it works, though - it's a bit of a noodle-baker!

It looks like we can just jump to "find the kth element in the original
list which re-parents nodeToReparent" (where k =
numOfReparentingThatReparentsNode, which is easily computed) - I'm
slightly perplexed as to why we can ignore whether reparentings have
been removed or not, though.