MVCN2TST,TGC: Fix bug with calcFinalDecryptionKey
In the original Editorial solution, fixParentChildAndHeights was called
in main() prior to calling calcFinalDecryptionKey.

In the testcase generator, I manually moved this call to
fixParentChildAndHeights into calcFinalDecryptionKey, but put it to
late.

It has now been moved to the first statement in calcFinalDecryptionKey.

It is, however, now giving another error.

UPDATE: Oh, silly me - the testcase isn't actually performing encryption
yet(!), so this obviously won't work at this stage!