MVCN2TST,TGC,REFACTOR: Begin to refactor so that dfs visits are separate
This is a step towards being able to compute dfsBegin/EndVisit
(necessary for Node::isDescendantOf to work) without calling
computeLookupInfo first.

Currently, because we are calling computeLookupInfo twice (once in the
testcase generator bit so that we can call isDescendantOf; the other in
setQueryIndexForQueries, where it should be called) which leads to RAM usage that
is too big for the Raspberry Pi to handle.