MVCNTST2,TGC: Snapshot of new testcase
This is one where a "squat" graph (approx 16 height) is created where
each node has degree at least three - the intention is that the range of
descendants of each node is quite variable, from small numbers (for the
 nodes nearer the "bottom") to large-ish numbers for nodes closer to the root.

It was hoped that this would be a good stress test for Phase 3 (where we
choose the kth non-descendant of a node at a given height), but in
practice, the number of descendants of a node skews very low, and it
might be easy to brute-force a solution - brute force might actually
end up being faster than the current solution XD. Such a brute-force
solution could simply take the full set of nodes at that height and
remove the descendants of the nodeToReparent - if the number of
descendants always skews small, as it does, then will be very effective.

Will have to experiment more; one simple avenue of cheating, given the
small number of heights, is to (for each node v) cache the set of node ids
at a given height in sorted order that are not descendants of v - might
try a cheat solution that tries this approach later.