MVCN2TST: Add initial attempt at Cheat solution
This cheats on Phase 3 by maintaining a cache of the ordered_set
representing the set of non-descendants of a (nodeToReparent,
newParentHeight) pair when the number of descendants exceeds a given
cutoff.

If the cache is not used (i.e. the number of descendants is small) then
it simply removes the set of descendants from the tree representing the
set of nodes at the given height, finds the kth, then re-adds the
descendants.