MVCN2TST,TGC: Fix another source of non-determinism
The 2nd squat graph test had calls to rnd.next(...) for *both* arguments
to addNodeChain (the "arm root" and the "arm size", and the order in which these calls are evaluated are
compiler-dependent - this caused the Raspberry Pi to generate a
different testfile for this test.

Fix by pulling out the "arm root"s into separate variables, computed
before the statement that calls addNodeChain.