MVCN2TST,TGC: Give the latest squat graph tc have 200'000 queries
Some small-ish ones - that should be processed quickly by both the
Editorial and the CHEAT_PHASE_THREE solution - have been added to make
up the difference (there used to be just 150'000 queries for this
testcase).

This might be reverted, or at least fine-tuned, if the Editorial
solution now takes more than 1.5 seconds to process this on Codechef
servers.

It would be nice to have 200'000 queries whatever the case, though, as it
helps to muddy the waters if all test files have
maxQueriesOverAllTestcases queries.