MVCN2TST,EDC: Remove the nodeBlockSize "optimisation"
It complicates the code quite a bit (both inside AVLTree, but also at
all call-sites of AVLTree, where we have to explain why we're choosing
a particular nodeBlockSize), but worse than that, it doesn't seem to
provide much of a performance boost over just using a deque to hold the
m_nodes.  Like, at all - in fact, some experiments run *faster* now that
the "optimisation" has been removed!

This is presumably because deque has a sensible "nodeBlockSize" built
into it, and doesn't need all my clumsy attempts at implementing
another.