Add peekLine() method to TestFileReader
For e.g. Simple Markdown Editor, the number and types of items in the
line depend on the contents of the line e.g. a line beginning with "F"
with have two items (a char (the "F"), followed by an int64_t), but a line
beginning with an 'N' will have three (a char (the "N"), followed by two
int64_ts).

This causes problems with TestFileReader.readLine - it needs to know the
exact contents of the line, which we don't know before we've seen the
line!

As a workaround, allow us to "peek" at the next line that will be read.
This simply returns the next line to be read (L, say), and then seeks back so
that the next call to readLine will again return the same L.

Simple Markdown Editor can now peek at the line; check the first
character; and then call readLine on that same line once we've
determined how many items there are on that line.