SIMKDOWN,TGC: Begin moving towards saner positions for queries/ inserts
At the moment, choosing a random position disproportionately chooses
positions in the middle of a long range of non-formatted characters -
and of course, if we then insert *more* non-formatted characters at this
position, then this range of non-formatted characters is even more
likely to have non-formatted characters inserted in it in the future,
resulting in a distinctly non-random (possibly exploitable?) distribution of inserts
(and queries).

It would be better to instead choose one of the formatted characters
with uniform probability and then pick a random position in the
run of non-formatted characters to its left.

For Queries, we need to pick a formatted char that has a non-zero run of
non-formatted to its left, so introduce AVLNode::totalFormattedDescendantsWithNonFormattedToLeft
and keep it updated.