SIMKDOWN,TGC: Begin moving towards better way to choose insertionPos
For inserting non-formatted chars.

As with the "queryPosition" case, choosing a random index in the
docuemnt will bias towards inserting in the middle of long runs of
non-formatting chars, and this bias is likely exploitable.

Better to choose a formatted char at random, then pick an insertion
point in the  run of non-formatted chars to its left - this will be more
"even".

However, I would also like to bias this towards adding
non-formatting chars to the left of formatting chars that *do not* yet
have a run of non-formatting chars to their left, so they do't get
"neglected" - this is what numFormattedCharsWithoutNonFormattingToLeft
is the beginning of work on this.