SIMKDOWN,TGC: Add first subtask2
This has maxNumTestcases testcases, each ending up creating a doc of
approx length subtask2.maxDocLength, and taking exactly
subtask2.maxQueriesPerTestcase queries.

Unfortunately, this is very slow to generate, mainly due to the call to
chooseRandomValuesWithSum2 for distributing the numNonformattedCharToAddForQuery across the
InsertNonFormatting queries so that it the doc ends up with maxDocLength chars in total - since
chooseRandomValuesWithSum2 is O(numToChooseFrom) and maxDocLength is
~1e9, this takes many seconds per testcase.

Should probably cache the results in a file.