Add a means of saving/ restoring a random_t object
Makes caching a bit easier - using the cached version vs not using the
cached version will lead to a different state for rnd, and so it will
generate different numbers depending on whether we use the cache or not,
which is unacceptable.

This allows us to eliminate this problem - when saving the cached
results after computing them, save the state of rnd; when instead we
load the cached version instead of doing the computation, load and
restore the state of rnd.  Then, whichever of the two paths we choose,
we end up generating the same list of random numbers afterwards.