Add quick checker for e.g chooseRandomValuesWithSum
chooseRandomValuesWithSum2 is very slow, so I wondered if there was a
better way - perhaps that given in chooseKWithSumExperimental.  So I
wrote a quick checker to test uniformity.

Turns out, though, that while chooseRandomValuesWithSum/
chooseRandomValuesWithSum2 given a nice, uniform distribution of choices
according to the checker, chooseKWithSumExperimental is hopelessly
lopsided.

There *might* be a way of fixing it, though - will check some other
time.