Fix assertion in chooseKRandomIndicesFrom
When verifying that all chosen indices are distinct, they are placed
into a set and the size of the set tested against the size of the
vector, chosenIndices.

However, the type of this set is set<int>, whereas the type of
chosenIndices is vector<int64_t>.

Change the size of this set to set<int64_t> to match chosenIndices.

This fixes a bug where the assertion was tripped erroneously, presumably
because two of the values in chosenIndices, *when cast from int64_t to
int*, had the same value.