Fix overflow bug in chooseKRandomIndicesFrom
The "value" of each AVLNode in the AVLTree "removedIndices" is a
(potentially large) index, and so maintaining a sum of
the values of all descendents of a node can easily overflow an int64_t.

Since we don't need this sum (it doesn't represent anything sensible,
anyway!), call AVLTree::setMaintainSumOfDescendentValues(false) to
disable it.