SIMKDOWN,EDC: Fix potential bug with reading in encrypted arguments
The type of encryptedArgument and encryptedArgument2 is int64_t, but we
were reading them as int - read them as int64_t instead.

Note that none of the testcases generated so far actually triggered this
bug - since the maxDocLength is <= 1e9, all of the encryptedArgument/
encryptedArgument2's are approximately this size (maybe as large as the
next power of 2, minus 1), and so fit into an int.  With a significantly
larger maxDocLength, though, this would almost certainly trigger a bug.