MVCN2TST,TGC: Add a "phase one correctness" test to tc 007
i.e. "the three long arms" one.

This tests for an off-by-one error in Phase One calculation: a correct
Phase One calculation requires the candidate to find the first
numCanReparentToPrefixSum that is *strictly greater than*
indexInOriginalList, but originally (in test 007), finding the first that is *greater
than or equal to* it would also give the correct answer.

With this new test, which picks a random nodeToReparent and finds the
*first (in the usual order) newParentNode that it can reparent to*,
choosing the numCanReparentToPrefixSum that is equal to
indexInOriginalList will give the wrong answer.

Note that some of the "large number of testcases" tests *did* correctly
guard against this incorrect approach, but I wanted one of the "large
graph" testcases to also guard against it, in case candidates used a
brute-force approach for smaller graphs.