MVCN2TST,TGC: Slightly refactor 007 "reparent to first possible node"
i.e. the query that stress-tests Phase One.

This now generates a different testcase, so not really a "refactoring",
but the generated testcase still fulfills all requirements.

The refactoring merely re-positions the "manually scramble all node ids"
bit to somewhere less "magical" and then manually re-calls
computeDFSInfo so that "isDescendantOf" works.

The prior revision required the scrambling to be done at a very specific
point (with no clue given as to why it had to be done precisely there)
and was error-prone; this new way is less magical and better-explained.