MVCN2TST,TGC: For 007, use new TreeGenerator::scrambleAndwriteTestcaseNG
This makes things a little bit less "magical" - all we need to do now is
use the new scrambleAndwriteTestcaseNG instead of scrambleAndwriteTestcase,
instead of having to manually scramble the queries and edges.

I *might* look at the impact of just deleting the old "scrambleAndwriteTestcase"
and replacing it with scrambleAndwriteTestcaseNG (i.e. for *all*
testcases, not just 007), but I think maybe I won't - don't want to rock
the boat, too much :)