MVCN2TST,TGC: Rename confusing variable in addWeightedQueries
The "chooseLowerNode" boolean, if true, means "choose a new parent whose
height [i.e. - distance from root] is larger than nodeToReparent", whereas the name kind of suggests
the opposite, to me.  Probably a side-effect of the general problem of
using the word "height" instead of, say, "distance".

Rename it to the clunkier - but more accurate -
"chooseNodeWithLargerHeight".