MOVCOIN2,TGC: Add another tc based on "3 long arms" but with 50% hasCoin
I've been slightly worried that the "number of nodes that hasCoin is
7%-85%" theme of the existing testcases is exploitable in some way, so
here's a large testcase where approx 50% of nodes hasCoin.

It's almost identical to the earlier "Three long arms" testcase, just
with a different seed and a random percentage (in the range 45%-55%) of nodes that hasCoin.

Might tweak the seed a little more.