MOVCOIN2,PS: In .odg files containing "scissors" svg, flatten svg to curves
When exporting an .odg file that has had the "scissors.svg" pasted into
it, exporting the .odg as .svg results in an svg where the scissors are
vertically stretched and look weird as hell.

Flattening the "scissors" subobject into curves fixes this and allows an
accurate export .svg.

The updated .png files have been included, though they are not really
changed.

The "pencil" subobject has a similar problem and will be fixed shortly.