MOVCOIN2,PS: Flattern "pencil" svg to curves in all .odg files
As with the "scissors" case, in .odg files where the pencil.svg was
pasted in directly as a .svg, exporting the .odg to .svg results in an
.svg with a distorted (vertically stretched) pencil.

Converting the pencil.svg subobject to curves, as done here in all .odg
files the contain the pencil.svg, fixes this problem.

Also includes the updated .pngs, though these should be almost identical
to the previous revision.