MVCN2TST,PS: Flattern "pencil" & "scissors" svg to curves in all .odg files
As mentioned in earlier commits, with .odg files where the pencil.svg
and scissors.svg were pasted in to Libre Office directly, exporting the
.odg to .svg gives an .svg where the scissor and pencil are distorted
(they are stretched vertically).

Fix this by reducing them to curves instead of svgs.

Also includes exported .png, though this should be virtually identical
to the previous version.