MOVCOIN2: Abandoning it as brute-force solution beats it
See tncks0121-36783124.cpp, introduced in the previous commit.

I can't see any way of toughening up the testcases to beat the
brute-force solution - the testcases were already about as hard as I
could make them.

Still bewildered as to how that solution runs so quickly, but it does,
so retiring MOVCOIN2.

In SEPT20, the problem "Counters on a Tree" has essentially been renamed
to MOVCOIN2, and so is taking this original MOVCOIN2's name and place.