COUNTREE: Add an (incomplete) cheat solution that breaks things :/
It currently gives the wrong answer, but should be representative of a
solution that can be made correct.

It breaks things because it revolves around evaluating the expression:

  grundyNumberForNode[j] ^= (distance + distOfNode[j]);

which, like a similar expression in the original "Move the Coins", turns
out to be evaluatable approx 5x10^9 times per second(!) on the Codechef
servers!

Needs tidying up.

Could conceivably be broken by upping the Node limit to 300'000 and
correspondingly upping the time limit to 7 seconds (and discarding/ condensing some of the crappier
testcases) - in that case, will probably take ~13 seconds to run,
without much room to optimise further.  Dunno, though.

Have notified the Codechef guys that this Problem *again* is probably
broken - I doubt I'm very popular there, right now XD

Very, very surprised (and disappointed :() right now.