COUNTREE,TGC: Increase the node limit to 300'000
The original limit of 200'000 left us very vulnerable to
counters-on-a-tree-cheat.cpp (especially with gcc 9.1) which compiles
down to code that runs so fast that an O(N^2) approach is only slightly
slower than an O(N * log N * log N) one.

This increase means that we will have to increase the time limit from 3s
to - say - 7s (or thereabouts), and with 17(!) testfiles this will put a big
strain on the Codechef servers, so in the following commits, I'll be
pruning out all the sillier ones.