COUNTREE,EDA: Fix glitch with DistTracker grundy number update
The "scroll digit" animation glitched as the
grundy_value_mobject.digitValue did not have an updated value: so e.g.
while it might appear that it's current value was 3, when scrolling it
to 5, it would instantly revert to an out-of-date value (1, say) and
*then* scroll from 1 to 5.

Update the grundy_value_mobject.digitValue when we addToAllDist and its visual
representation changes.