compile-latest-cpp.sh: Fix bug (compile errors reported as Successful!)
The call to:

    set +x

clobbered the exit code value ($?) for the actual compilation and always
set it to 0 ("Successful").

Use a different approach from using "set +/-x": just form the command to
run as a string; print that string; then execute it; no calls to "set
+/-x" required.

This fixes that bug and makes the output less noisy, too.