AoC 2020 Day 25: Improve type safety in calcLoopSize
It returns an int64_t, so the variable ("loopSize") it returns should be of that
type, too, rather than "int" like it was before!

Doesn't actually make any difference in this case (Fermat's Little
Theorem tells us loopSize will never exceed 20201227, which fits
comfortably into an int), but let's not be sloppy! :p