AoC2019: Add "takeOutput" convenience method to IntCodeComputer
This combines "output()" and "clearOutput()" into one function, since:

* We almost always want to do both at once
* It's easy to forget to clear the output, which has caused subtle bugs
  in the past!

The only reason I didn't add this originally is because it could not be
implemented in an exception-safe way (the vector *copy* constructor could throw
*after* we cleared m_output, in which case we don't have
"commit-or-rollback" semantics when exceptions are thrown), but in
C++11, which we're using, the returned vector will be *moved* instead of
*copied*, and vector's move constructor, like most move constructors, is
noexcept, so this cannot happen.  Woo!