AoC2019D16P2: REFACTOR: Simplify code around computing outputVal
e.g. countdownToNextBase is (unlike in Part 1) nonsensical, here: we
don't do any counting down!

Introduce "blockSize" instead, which serves a similar purpose but fits
in better with the explanatory comment, which takes about adding
"blocks".  Generally, re-arrange the flow so that it is less Rube
Goldberg-esque.

Also, add a bit more documentation.