AoC19D17P2: WIP - add buildValidCommandListForAllStates
As the name suggests, this builds the list of valid lists of commands
for each state (Coord + starting direction).

Currently, there can be 10's of thousands(!) of command lists for a
given state, though most are degenerate i.e. just meander around, and
mostly backtrack.

This is too much to comfortably deal with, so we'll need some pruning:
I'm not sure if we can get it down to a reasonable number, though :/

Some redundancies have already been omitted e.g. never move forward
twice in a row; never turn twice in a row; etc.