AoC19D17P2: correct the calculation of command-list string lengths
The old one was wildly wrong; the encoding of:

    left, left, forward 3, right

should be:

    L,L,3,R

but I erroneously calculated it as:

    L1,L1,F3,R

or maybe as:

    L2,F3,R

Idiot XD

The previous encoding implicitly prohibited e.g. 3 Ls in a row - I've
made this prohibition explicit, now.

Happily, the effect of this is to reduce the number of valid commands
lists per state from 10's of thousands to just (usually) thousands.