AOC19D22P2: begin to replace __uint128_t with ModNum class
The ModNum class (which I wrote ages ago for some Hackerrank challenges)
is designed to cut out some of the drudgery/ error-prone-ness of dealing
with modular arithmetic.

This version has been modified to use bitwise multiplication as ordinary
multiplication with a modulus this large (119315717514047(!)) will often
overflow an int64_t.

For now, the __uint128_t implementation remains, and we merely assert
that the ModNum gives identical results to it - this has been tested
with all sample inputs for this Problem, plus the main input.