AoC22D16P1: optimise by not storing "in transit" states
With this commit, when doing the state exploration, we no longer always
explore all neigbouring valves: instead, we pick a target valve with
non-zero flow; look up the distance to this target valve; and only store
the state in which we've reached this target valve

Thus, we now skip the states in which we're wandering towards the target
(which might include pointless dithering/ backtracking).

This speeds up finding the solution by approx 5x.

I suspect we can do even better, though, by removing pressureReleased
from the State and storing only the "best" pressureReleased for these
(truncated) States - will give that a try; if all goes well, you'll see
it in a subsequent commit :)