AoC22D16P2: WIP; broken - snapshot of experiment
Attempting to find a compromise between full implementation of "skip
in-transit positions" and the previous implementation.

This is not going well - not only does this approach give the wrong
answer for the sample problem, it also generates even more states than
before.  Sigh.

I don't know whether fixing the "wrong answer" will fix the state-space
explosion - need to think more.